requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.1
python-dotenv==1.0.0
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8
lxml==4.9.3